Impementation Modulation & refactor
