+++ .idea/compiler.xml
@@ -1,6 +1,8 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="CompilerConfiguration">
-    <bytecodeTargetLevel target="11" />
+    <bytecodeTargetLevel target="1.8">
+      <module name="YelpApp.app" target="11" />
+    </bytecodeTargetLevel>
   </component>
 </project>
