+++ app/build.gradle
@@ -40,6 +40,10 @@ android {
 }
 
 dependencies {
+    //MODULATION
+    implementation project(":domain")
+    implementation project(":usecases")
+    implementation project(":data")
 
     implementation 'androidx.core:core-ktx:1.7.0'
     implementation 'androidx.appcompat:appcompat:1.4.1'
@@ -74,7 +78,7 @@ dependencies {
     implementation "androidx.activity:activity-ktx:1.4.0"
     implementation "androidx.fragment:fragment-ktx:1.4.1"
     //PlayServices
-    implementation 'com.google.android.gms:play-services-location:17.1.0'
+    implementation 'com.google.android.gms:play-services-location:19.0.1'
     //rxjava for observables
     implementation 'io.reactivex.rxjava2:rxandroid:2.1.1'
