--- app/src/main/java/com/example/yelpapp/data/repository/SearchBusinessesRepository.kt
@@ -1,9 +1,0 @@
-package com.example.yelpapp.data.repository
-
-import com.example.yelpapp.data.entity.YelpSearchResult
-import io.reactivex.Observable
-
-
-interface SearchBusinessesRepository {
-    fun getSearchBussiness(): Observable<YelpSearchResult>
-}
