--- app/src/main/java/com/example/yelpapp/data/repository/SearchBusinessesRepositoryImp.kt
@@ -1,4 +1,0 @@
-package com.example.yelpapp.data.repository
-
-class SearchBusinessesRepositoryImp {
-}
