+++ build.gradle
@@ -2,8 +2,9 @@
 plugins {
     id 'com.android.application' version '7.1.3' apply false
     id 'com.android.library' version '7.1.3' apply false
-    id 'org.jetbrains.kotlin.android' version '1.6.10' apply false
+    id 'org.jetbrains.kotlin.android' version '1.6.21' apply false
     id 'androidx.navigation.safeargs.kotlin' version '2.4.1' apply false
+    id 'org.jetbrains.kotlin.jvm' version '1.6.21' apply false
 }
 
 task clean(type: Delete) {
