+++ data/src/main/java/com/example/yelpapp/data/entity/Coordinates.kt
