+++ data/src/main/java/com/example/yelpapp/data/entity/YelpSearchResult.kt
