+++ usecases/src/main/java/com/example/yelpapp/usecases/MyClass.kt
@@ -1,0 +1,4 @@
+package com.example.yelpapp.usecases
+
+class MyClass {
+}
